Remove special handling of default implementations in Task
The special handling has been introduced since embulk-core v0.9.0 in:
https://github.com/embulk/embulk/pull/889

Its goal was to deprecate Joda-Time's DateTimeZone used in some of embulk-core's
predefined Task, with compatibilities kept.

The embulk-util-config library is totally for plugins, on the other hand.
Everything should be on plugin's side, and each plugin does not need to
consider compatibility just in itself.